requirements.txt
python-fasthtml==0.3.2